clean up after reject
Implemented to move the trial-channel to declined category.

Also now everything after the mention is the reason for the decline.
